+++ src/document/printers/directivePrinter.ts
@@ -9,6 +9,7 @@ import { TransformOptions } from '../transformOptions';
 import { ArrayPrinter } from './arrayPrinter';
 import { IndentLevel } from './indentLevel';
 import { ParserOptions } from "prettier";
+import { getPrintWidth } from './printWidthUtils';
 
 export class DirectivePrinter {
     private static defaultControlDirectiveNames: string[] = [
@@ -181,22 +182,3 @@ function removeContentLines(content: string): string {
     return newContent;
 }
 
-function getPrintWidth(content:string, defaultWidth:number): number {
-    if (content.includes("\n") && (
-        content.includes('match') || content.includes('=>' || content.includes('&&')
-    ))) {
-        const lines = StringUtilities.breakByNewLine(content);
-        let maxW = 0;
-
-        lines.forEach((line) => {
-            const lineL = line.trim().length;
-            if (lineL > maxW) {
-                maxW = lineL;
-            }
-        });
-
-        return maxW + Math.ceil(maxW / 2);
-    }
-
-    return defaultWidth;
-}