+++ src/document/transformOptions.ts
@@ -1,5 +1,3 @@
-import { ParserOptions } from "prettier";
-
 export interface TransformOptions {
     spacesAfterDirective: number,
     spacesAfterControlDirective: number,
@@ -7,5 +5,6 @@ export interface TransformOptions {
     formatDirectivePhpParameters: boolean,
     formatDirectiveJsonParameters: boolean,
     formatInsideEcho: boolean,
-    phpOptions: any|undefined
+    phpOptions: any|undefined,
+    echoStyle: string
 }