+++ src/test/formatter_php_blocks.test.ts
@@ -138,4 +138,29 @@ fn () => true;
 `;
         assert.strictEqual(formatBladeString(template), out);
     });
+
+    test('it is smart about print width', () => {
+        const input = `
+@php
+    $buttonClasses = \\Illuminate\\Support\\Arr::toCssClasses([
+        'filament-tabs-item flex items-center h-8 px-5 font-medium rounded-lg outline-none focus:ring-2 focus:ring-primary-500 focus:ring-inset',
+        'hover:text-gray-800 focus:text-primary-600' => ! $active,
+        'dark:text-gray-400 dark:hover:text-gray-300 dark:focus:text-primary-400' => (! $active) && config('filament.dark_mode'),
+        'text-primary-600 shadow bg-white' => $active,
+    ]);
+@endphp
+`;
+        const out = `@php
+    $buttonClasses = \\Illuminate\\Support\\Arr::toCssClasses([
+        "filament-tabs-item flex items-center h-8 px-5 font-medium rounded-lg outline-none focus:ring-2 focus:ring-primary-500 focus:ring-inset",
+        "hover:text-gray-800 focus:text-primary-600" => ! $active,
+        "dark:text-gray-400 dark:hover:text-gray-300 dark:focus:text-primary-400" => ! $active && config("filament.dark_mode"),
+        "text-primary-600 shadow bg-white" => $active,
+    ]);
+@endphp
+`;
+
+        assert.strictEqual(formatBladeString(input), out);
+        assert.strictEqual(formatBladeString(out), out);
+    });
 });